SEPARATE OUT MATCHERS
Since all the code for matchers happens to be in main.py thus making it quite cluttered therefore a new file matchers.py has all the code for matchers for which has appropriate function calls in main.py.
